Fixed broken single-precision compilation
git-svn-id: https://svn.berlios.de/svnroot/repos/glimmer-cism/glimmer-cism/glimmer/branches/RELEASE_1_0_0_BRANCH@804 2cfc7cc4-626f-0410-8325-f6cb4dace9bf

+++ src/fortran/glint_mpinterp.F90
@@ -275,8 +275,8 @@ contains
   subroutine mean_preserve_interp(params,in,out,zeros)
 
     type(mpinterp),         intent(in)  :: params
-    real(dp),dimension(:,:),intent(in)  :: in
-    real(dp),dimension(:,:),intent(out) :: out
+    real(rk),dimension(:,:),intent(in)  :: in
+    real(rk),dimension(:,:),intent(out) :: out
     logical, dimension(:,:),intent(out) :: zeros
     
     integer :: i,j,ii
